drm/exynos: hdmi: Implement initialize op for hdmi
This patch implements the initialize callback in the hdmi and mixer
manager. This allows us to get rid of drm_dev in the drm_hdmi level and
track it in the mixer and hdmi drivers. This is one of the things
holding back the complete removal of the drm_hdmi layer.

Change-Id: I56a3e31a0fc79693bc486d0ff4ba146ca74d783b
Signed-off-by: Sean Paul <seanpaul@chromium.org>