drm/exynos: Rename display_op power_on to dpms
This patch renames the display_op power_on to dpms to accurately reflect
what the function does.

The side-effect of this patch is that the new hdmi dpms callback is now
invoked twice in the dpms path. This is safe and will be dealt with when
the exynos_drm shim goes away.

Change-Id: I49d2ad045a8204d7a669ecb58b22811b12153b6a
Signed-off-by: Sean Paul <seanpaul@chromium.org>