drm/exynos: Clean up FIMD power on/off routines
This patch separates the fimd_activate function into poweron/poweroff
functions to be more consistent with the other drivers in exynos drm. It
also properly cleans up after failures in poweron. The functions have
also been shuffled around such that they are all in the same
spot in the file and poweron/poweroff can be called from the dpms function.

Change-Id: I3a4f242559a6f3f1b2f634450fdbbd196231a790
Signed-off-by: Sean Paul <seanpaul@chromium.org>